requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==1.0.2
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
lyricsgenius==3.0.1
MarkupSafe==2.0.1
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
tekore==3.7.1
telegraph==1.4.1
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.4.1
